requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.1
asyncpg==0.26.0
atlastk==0.13.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dictdiffer==0.9.0
fastapi==0.79.0
flower==1.1.0
h11==0.13.0
humanize==4.2.3
idna==3.3
iso8601==1.0.2
kombu==5.2.4
pika==1.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pydantic==1.9.1
pypika-tortoise==0.1.6
pytz==2022.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomlkit==0.11.1
tornado==6.2
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
